version-bridge: restructure module layout
