+++ packages/version-bridge/src/lib/VersionedContext.ts
@@ -16,31 +16,7 @@
 
 import { createContext, useContext, Context } from 'react';
 import { getGlobalSingleton, setGlobalSingleton } from './globalObject';
-
-/**
- * The versioned value interface is a container for a set of values that
- * can be looked up by version. It is intended to be used as a container
- * for values that can be versioned independently of package versions.
- */
-export type VersionedValue<Versions extends { [version: number]: any }> = {
-  atVersion<Version extends keyof Versions>(
-    version: Version,
-  ): Versions[Version] | undefined;
-};
-
-/**
- * Creates a container for a map of versioned values that implements VersionedValue.
- */
-export function createVersionedValueMap<
-  Versions extends { [version: number]: any },
->(versions: Versions): VersionedValue<Versions> {
-  Object.freeze(versions);
-  return {
-    atVersion(version) {
-      return versions[version];
-    },
-  };
-}
+import { createVersionedValueMap, VersionedValue } from './VersionedValue';
 
 export function useVersionedContext<
   Versions extends { [version in number]: any },
